+++ app/index.js
@@ -55,6 +55,8 @@ const {
     defaultUseragent
 } = require("./util")
 const hljs = require("highlight.js")
+const {JSDOM} = require("jsdom")
+const {Readability} = require("@mozilla/readability")
 const {"sync": rimrafSync} = require("rimraf")
 const rimraf = pattern => {
     try {
@@ -691,8 +693,7 @@ let blocker = null
 let permissions = {}
 const sessionList = []
 const adblockerPreload = require.resolve("@cliqz/adblocker-electron-preload")
-const hlStyling = readFile(joinPath(__dirname,
-    "../node_modules/highlight.js/styles/github-dark.css"))
+const defaultCss = readFile(joinPath(__dirname, `colors/default.css`))
 ipcMain.on("set-redirects", (_, rdr) => {
     redirects = rdr
 })
@@ -942,9 +943,30 @@ ipcMain.on("create-session", (_, name, adblock, cache) => {
             let body = ""
             res.on("end", () => {
                 const hl = hljs.highlightAuto(body)
-                call(`<!DOCTPYE html>\n<html><head><style>${hlStyling}</style>
+                call(`<!DOCTPYE html>\n<html><head><style>${defaultCss}</style>
                 <title>${encodeURI(req.url)}</title>
-                </head><body><pre><code>${hl.value}</code></pre></body></html>`)
+                </head><body id="sourceviewer">
+                    <pre><code>${hl.value}</code></pre></body></html>`)
+            })
+            res.on("data", chunk => {
+                body += chunk
+            })
+        })
+        request.on("abort", () => call(""))
+        request.on("error", () => call(""))
+        request.end()
+    })
+    newSession.protocol.registerStringProtocol("readerview", (req, call) => {
+        const url = req.url.replace(/readerview:\/?\/?/g, "https://")
+        const request = net.request({"partition": name, url})
+        request.on("response", res => {
+            let body = ""
+            res.on("end", () => {
+                const dom = new JSDOM(body, {url})
+                const out = new Readability(dom.window.document).parse().content
+                call(`<!DOCTPYE html>\n<html><head><style>${defaultCss}</style>
+                <title>${encodeURI(req.url)}</title>
+                </head><body id="readerview">${out}</body></html>`)
             })
             res.on("data", chunk => {
                 body += chunk